Move criteria query building to filter config
